chore(bors): do not build 'master'
Because bors now controls the merge process, it will always run a build
on `staging` before fast-forwarding `master` to that point, meaning the
commit that is pushed will already have a green build light on it.
Adding 'master' just means the build will get run twice for a single
commit, which could get expensive.

Signed-off-by: Austin Seipp <aseipp@pobox.com>