fix: use `read_verilog -defer` for prims
By default, when the Yosys Verilog frontend attempts to read Verilog, it
does so too 'eagerly'. For instance, given:

    module t(input wire clk);
      parameter FILE = "";

      reg [31:0] rom[0:8191];
      initial begin; $readmemh(FILE, rom, 0, 8191);
    endmodule

Then any use of the module `t` with a parameter `FILE` will result in an
error:

    top.v:0: ERROR: Can not open file `` for \$readmemh.

i.e. when Verilog is parsed to the AST, the argument to `readmemh` is
evaluated at parse time, not at elaboration time during synthesis.

This can be fixed by using the `-defer` option to read_verilog, but the
design of this option means that:

 - The file containing the module with the above code (i.e. `t.v`) must
be read using `-defer`, not the calling module.

 - The AST read from read_verilog is "empty" at first and is only
elaborated later during a `hierarchy` pass, where the AST is elaborated
further into RTLIL.

Because this changes what cells get loaded into an `RTLIL::Design` once
you read the module, we need to slightly rework the loading logic to
handle this change. However, as a result, we can now load any of the
needed.

Two extra changes:

 - Rewrite the test to use a trivial BRAM module from the Prelude, which
triggers the above code (via `BRAM2Load.v` in bsc.)

 - Error if we use modules that Yosys can't yet support, notably
`ProbeHook` and `InoutConnect`.

Signed-off-by: Austin Seipp <aseipp@pobox.com>